requirements.txt
gunicorn
psycopg2
BeautifulSoup
pygeoip
south
dnspython
